bcm53xx: Fix kernel v6.1 config
The v6.1 kernel does not build for me, building with -j1 V=s
reveals that it is prompting for a kernel config option.
I also needed to add CONFIG_MTD_SPLIT_FIRMWARE for splitting
to work.

This make the v6.1 config work nicely as a baseline for my
DIR-890L patches.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
---
ChangeLog ->v3:
- New patch to make kernel v6.1 work