modules: Add kernel module for MV88E6xxx DSA switch
This adds a kernel module for the Marvell MV88E6XXX DSA
switch.

We can't just compile this in because just a few devices
has this DSA, and it depends on e.g. the I2C and SFP
to be loaded as modules first.

Compile in the DSA tagger, it's small code and just a
hassle to turn it into a module.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>