mtdsplit_uimage: Split also after offsetted uImage
The uImage splitter recognizes a rootfs either:

1. Right after the uImage if it comes first in the
   partition or
2. Before the uImage if it is located at an offset
   inside the partition.

Add a third case:

3. After the uImage also at an offset inside the
   partition, if and only if 1 and 2 fails.

The reason why this is needed is because on the
BCM6328-based Inteno XG6846 we need to put a small
U-Boot binary first in the partition, then the uImage,
then the rootfs.

The U-Boot binary that comes first cannot be split off
into its own partition in this case because it needs
to be part of the bigger "firmware" partition. Which
we use for installation and upgrades.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>