Resurrect OpenWrt IXP4xx support
XP4xx was deleted because of lack of maintenance in 2020.

In the years since, the upstream Linux support for IXP4xx has
been rewritten from scratch. It is now pretty well supported
using device tree and modern subsystems that didn't exist
20 years ago when the first IXP4xx port was done.

With the switch to kernel v6.1, OpenWrt has the required
baseline to bring IXP4xx back.

With the (suspiciously similarly named!) IPQ4xxx now supporting
v6.1 exclusively, the door should be open for IXP4xx!

This is a reimplementation of the IXP4xx support which is
using all the contemporary abstractions, the only thing
actually resurrected is the APEX boot loader. The port is
thus "slim" and not patching a lot on the mainline kernel
as pretty much everything is upstream.

As a starter, we bring back two Gateworks reference designs,
one for IXP42x and one for IXP43x, and the classic NSLU2
NAS. We also bring back USRobotics USR8200 because it has
active users and testers.

Here you find binaries and illustrations if you want to check
out how the support works right now, if you have an NSLU2
you can flash the binary and rootfs and it "just works",
though NSLU2 is not a default build:
https://dflund.se/~triad/krad/ixp4xx/

Thanks to Imre Kaloz add Krzysztof Halasa for all cheering
and reference designs :)

Thanks to Howard Harte and Joseph for stepping in and helping
out with device support and debugging.

Thanks for Hauke for the the in-depth review of the v1
patch set. Thanks to Christian Marangi, Robert Marko and
Jonas Gorski for reviewing the v2 patch set.

I am pretty confident that we can bring back any IXP4xx with
this, they just need some time and focus. I will add more
routers as time permits.

I don't know who can import this from patches or pull request:
it may be best if I just apply for commit access to the
OpenWrt git repository and import it myself perhaps?

To: Imre Kaloz <kaloz@openwrt.org>
To: Krzysztof Halasa <khalasa@piap.pl>
To: Howard Harte <hharte@magicandroidapps.com>
To: openwrt-devel@lists.openwrt.org
Cc: Hauke Mehrtens <hauke@hauke-m.de>
Cc: Raylynn Knight <rayknight@me.com>
Cc: Eric <evil.function@proton.me>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

---
Changes in v4:
- Bring the AMD Phy online at preinit, since this may sit
  on a primary network interface.
- Split off some USR8200 patches to generic kernel patches and
  moved these to the front of the series so they can be applied.
- Drop applied patches.
- Link to v3: https://lore.kernel.org/r/20231012-ixp4xx-v3-0-e09c74712c0b@linaro.org

Changes in v3:
- Drop applied patch.
- Move APEX to build from a git URL (Ansuel)
- Add a build rule for lib-chacha20 for armeb
- Split microcode firmware resurrection to its own patch
- Use AutoProbe on RTC modules and crypto module
- Name devices with vendor_device tuples
- Fix network script for MAC to use just ip not ifconfig
- Add a separate patch for the USRobotics USR8200 router.
- Deferring system upgrade for later since it is a major
  implementation and can be done on top (I still plan to do it!)
- Pushed to github and no build complaints. I suppose no news is
  good news.
- Link to v2: https://lore.kernel.org/r/20230928-ixp4xx-v2-0-ce011eef8239@linaro.org

Changes in v2:
- See details in patches, most feedback is adressed.
- Link to v1: http://lists.openwrt.org/pipermail/openwrt-devel/2023-June/041193.html

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 4,
    "change-id": "20230928-ixp4xx-a393c84d9468",
    "prefixes": [],
    "history": {
      "v2": [
        "20230928-ixp4xx-v2-0-ce011eef8239@linaro.org"
      ],
      "v3": [
        "20231012-ixp4xx-v3-0-e09c74712c0b@linaro.org"
      ]
    }
  }
}