Enable USB serial console on the D-Link DNS-313
This patch set makes it possible to plug in a USB cable
into the device USB B receptacle on the DNS-313 and
it will appear as /dev/ttyACM0 on the host and there will
be a serial console on it, at 115200 baud.

This is useful for users who don't like to solder in
their device to get a serial console.

The custom inittab is a bit ugly, but I've seen that this
seems to be what other devices do, other ideas welcome.

To: openwrt-devel@lists.openwrt.org
To: Robert Marko <robimarko@gmail.com>
To: Hans Ulli Kroll <ulli.kroll@googlemail.com>
To: Sebastian Luft <sebastian.luft@gmail.com>
To: Christian Lamparter <chunkeey@gmail.com>
Cc: Corentin Labbe <clabbe@baylibre.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20231111-dns313-usb-serial-v1-0-6402033bcdfb@linaro.org



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20231111-dns313-usb-serial-b9ac47e3a27e",
    "prefixes": [],
    "history": {
      "v1": [
        "20231111-dns313-usb-serial-v1-0-6402033bcdfb@linaro.org"
      ]
    }
  }
}